Add files via upload
In this lecture I learned about Feature Engineering and Polynomial Regression.
Problem statement:-What if your features/data are non-linear or are combinations of features? For example, Housing prices do not tend to be linear with living area but penalize very small or very large houses resulting in the curves shown in the graphic above. How can we use the machinery of linear regression to fit this curve? Recall, the 'machinery' we have is the ability to modify the parameters  w,  𝐛 to 'fit' the equation to the training data. However, no amount of adjusting of  𝐰, 𝐛 will achieve a fit to a non-linear curve.

SO we tried Feature Engineering (Given y=x^2).
1.Firstly we tried with Linear curve i.e normal x. But it will obviously not work.
2. Secondly we went with X= x**2. Now it worked perfectly and the values of w and b after gradient descent also suggested that the feature engineered was correct.

Also I learned about Selecting features(i.e which features are appropriate and which are not).
suppose Y=x**2
and we took X = np.c_[x, x**2, x**3] (Engineered features).
after running gradient descent we found the values of w tobe  [0.08 0.54 0.03] and b is 0.0106.This implies the model after fitting/training is:
0.08𝑥 + 0.54𝑥^2 + 0.03𝑥^3 + 0.0106.
Gradient descent has emphasized the data that is the best fit to the  𝑥2 data by increasing the  𝑤1 term relative to the others. If you were to run for a very long time, it would continue to reduce the impact of the other terms.
Gradient descent is picking the 'correct' features for us by emphasizing its associated parameter.

Also I learned about Complex functions.
Eg- y = np.cos(x/2). Now how will we fit our model if we dont know the function.
Here engineered features will take place i.e. X = np.c_[x, x**2, x**3,x**4, x**5, x**6, x**7, x**8, x**9, x**10, x**11, x**12, x**13].