Add files via upload
In this lab I learned about difference in regression and classification.

Examples of classification problems are things like: identifying email as Spam or Not Spam or determining if a tumor is malignant or benign. In particular, these are examples of binary classification where there are two possible outcomes. Outcomes can be described in pairs of 'positive'/'negative' such as 'yes'/'no, 'true'/'false' or '1'/'0'.

the lab demonstrated that the linear model is insufficient to model categorical data and therefore we need classification to solve categorical data.